+++ test/CreateHandler/CreateHandlerCommandTest.php
@@ -40,11 +40,16 @@ class CreateHandlerCommandTest extends TestCase
         $this->output = $this->prophesize(ConsoleOutputInterface::class);
 
         $this->command = new CreateHandlerCommand('handler:create');
+    }
 
-        // Do not require the generated handler during testing
-        $r = new ReflectionProperty($this->command, 'requireHandlerBeforeGeneratingFactory');
+    /**
+     * Allows disabling of the `require` statement in the command class when testing.
+     */
+    private function disableRequireHandlerDirective(CreateHandlerCommand $command) : void
+    {
+        $r = new ReflectionProperty($command, 'requireHandlerBeforeGeneratingFactory');
         $r->setAccessible(true);
-        $r->setValue($this->command, false);
+        $r->setValue($command, false);
     }
 
     private function reflectExecuteMethod()
@@ -57,17 +62,17 @@ class CreateHandlerCommandTest extends TestCase
     /**
      * @return ObjectProphecy|Application
      */
-    private function mockApplication()
+    private function mockApplication(string $forService = 'Foo\TestHandler')
     {
         $helperSet = $this->prophesize(HelperSet::class)->reveal();
 
         $factoryCommand = $this->prophesize(Command::class);
         $factoryCommand
             ->run(
-                Argument::that(function ($input) {
+                Argument::that(function ($input) use ($forService) {
                     Assert::assertInstanceOf(ArrayInput::class, $input);
                     Assert::assertContains('factory:create', (string) $input);
-                    Assert::assertContains('Foo\TestHandler', (string) $input);
+                    Assert::assertContains($forService, (string) $input);
                     return $input;
                 }),
                 $this->output->reveal()
@@ -83,21 +88,45 @@ class CreateHandlerCommandTest extends TestCase
 
     public function testConfigureSetsExpectedDescription()
     {
-        $this->assertContains('Create a PSR-15 request handler', $this->command->getDescription());
+        $this->assertContains(CreateHandlerCommand::HELP_HANDLER_DESCRIPTION, $this->command->getDescription());
+    }
+
+    public function testConfigureSetsExpectedDescriptionWhenRequestingAnAction()
+    {
+        $command = new CreateHandlerCommand('action:create');
+        $this->assertContains(CreateHandlerCommand::HELP_ACTION_DESCRIPTION, $command->getDescription());
     }
 
     public function testConfigureSetsExpectedHelp()
     {
-        $this->assertEquals(CreateHandlerCommand::HELP, $this->command->getHelp());
+        $this->assertEquals(CreateHandlerCommand::HELP_HANDLER, $this->command->getHelp());
+    }
+
+    public function testConfigureSetsExpectedHelpWhenRequestingAnAction()
+    {
+        $command = new CreateHandlerCommand('action:create');
+        $this->assertEquals(CreateHandlerCommand::HELP_ACTION, $command->getHelp());
     }
 
     public function testConfigureSetsExpectedArguments()
     {
         $definition = $this->command->getDefinition();
         $this->assertTrue($definition->hasArgument('handler'));
+        $this->assertFalse($definition->hasArgument('action'));
         $argument = $definition->getArgument('handler');
         $this->assertTrue($argument->isRequired());
-        $this->assertEquals(CreateHandlerCommand::HELP_ARG_HANDLER, $argument->getDescription());
+        $this->assertEquals(CreateHandlerCommand::HELP_HANDLER_ARG_HANDLER, $argument->getDescription());
+    }
+
+    public function testConfigureSetsExpectedArgumentsWhenRequestingAnAction()
+    {
+        $command = new CreateHandlerCommand('action:create');
+        $definition = $command->getDefinition();
+        $this->assertTrue($definition->hasArgument('action'));
+        $this->assertFalse($definition->hasArgument('handler'));
+        $argument = $definition->getArgument('action');
+        $this->assertTrue($argument->isRequired());
+        $this->assertEquals(CreateHandlerCommand::HELP_ACTION_ARG_ACTION, $argument->getDescription());
     }
 
     public function testConfigureSetsExpectedOptions()
@@ -107,16 +136,33 @@ class CreateHandlerCommandTest extends TestCase
         $this->assertTrue($definition->hasOption('no-factory'));
         $option = $definition->getOption('no-factory');
         $this->assertFalse($option->acceptValue());
-        $this->assertEquals(CreateHandlerCommand::HELP_OPT_NO_FACTORY, $option->getDescription());
+        $this->assertEquals(CreateHandlerCommand::HELP_HANDLER_OPT_NO_FACTORY, $option->getDescription());
 
         $this->assertTrue($definition->hasOption('no-register'));
         $option = $definition->getOption('no-register');
         $this->assertFalse($option->acceptValue());
-        $this->assertEquals(CreateHandlerCommand::HELP_OPT_NO_REGISTER, $option->getDescription());
+        $this->assertEquals(CreateHandlerCommand::HELP_HANDLER_OPT_NO_REGISTER, $option->getDescription());
+    }
+
+    public function testConfigureSetsExpectedOptionsWhenRequestingAnAction()
+    {
+        $command = new CreateHandlerCommand('action:create');
+        $definition = $command->getDefinition();
+
+        $this->assertTrue($definition->hasOption('no-factory'));
+        $option = $definition->getOption('no-factory');
+        $this->assertFalse($option->acceptValue());
+        $this->assertEquals(CreateHandlerCommand::HELP_ACTION_OPT_NO_FACTORY, $option->getDescription());
+
+        $this->assertTrue($definition->hasOption('no-register'));
+        $option = $definition->getOption('no-register');
+        $this->assertFalse($option->acceptValue());
+        $this->assertEquals(CreateHandlerCommand::HELP_ACTION_OPT_NO_REGISTER, $option->getDescription());
     }
 
     public function testSuccessfulExecutionEmitsExpectedMessages()
     {
+        $this->disableRequireHandlerDirective($this->command);
         $this->command->setApplication($this->mockApplication()->reveal());
 
         $generator = Mockery::mock('overload:' . CreateHandler::class);
@@ -142,6 +188,40 @@ class CreateHandlerCommandTest extends TestCase
 
         $this->assertSame(0, $method->invoke(
             $this->command,
+            $this->input->reveal(),
+            $this->output->reveal()
+        ));
+    }
+
+    public function testSuccessfulExecutionEmitsExpectedMessagesWhenRequestingAnAction()
+    {
+        $command = new CreateHandlerCommand('action:create');
+        $this->disableRequireHandlerDirective($command);
+        $command->setApplication($this->mockApplication('Foo\TestAction')->reveal());
+
+        $generator = Mockery::mock('overload:' . CreateHandler::class);
+        $generator->shouldReceive('process')
+            ->once()
+            ->with('Foo\TestAction')
+            ->andReturn(__DIR__);
+
+        $this->input->getArgument('action')->willReturn('Foo\TestAction');
+        $this->input->getOption('no-factory')->willReturn(false);
+        $this->input->getOption('no-register')->willReturn(false);
+        $this->output
+            ->writeln(Argument::containingString('Creating action Foo\TestAction'))
+            ->shouldBeCalled();
+        $this->output
+            ->writeln(Argument::containingString('Success'))
+            ->shouldBeCalled();
+        $this->output
+            ->writeln(Argument::containingString('Created class Foo\TestAction, in file ' . __DIR__))
+            ->shouldBeCalled();
+
+        $method = $this->reflectExecuteMethod();
+
+        $this->assertSame(0, $method->invoke(
+            $command,
             $this->input->reveal(),
             $this->output->reveal()
         ));